Working on the QueryBuilder abstract interface, #305. Refactored the code into general implementation (for user) and backend specific implementation. There is no more inheritance, the general implementation chooses a backend specific implementation as a member upon instantiation. An abstract class provides an interface, but more documentation and more checks are needed.
